Modifikime ne Program

+++ src/main/java/com/example/demo/service/JsPytjeService.java
@@ -1,9 +1,7 @@
 package com.example.demo.service;
 
-import com.example.demo.exceptions.KuiziException;
 import com.example.demo.exceptions.NashornException;
 import com.example.demo.model.JsPytjeEntity;
-import com.example.demo.model.KuizeTeLuajturaEntity;
 
 import java.util.List;
 
@@ -22,6 +20,6 @@ public interface JsPytjeService {
 
     List<String> findByPergjigja() throws NashornException;
 
-    List<JsPytjeEntity> findByQuizID(Long id) throws NashornException;
+    List<JsPytjeEntity> findByQuizID(Integer id) throws NashornException;
 }
